add GrammarChecker struct
